[docs] added some status informations
Added reports from goreportcard and travis-ci for the yummy project to the
README.md file.

goreportcard.com provides a service if gofmt and companions do not complain
about the project like the spelling error that was corrected before.

travis-ci.org is a great service that automatically runs checks on every commit
with various go versions and settings. Currently its a minimal setup.